Fix device index parsing in accelerator (#2328) (#807)
Currently, if the current device is passed as a string, e.g., "hpu:0",
the `current_device_index` is also set to the same string instead of the
actual device index. This commit updates the parsing code to actually
extract the device index from the string. Otherwise, the behaviour is
left untouched.

Co-authored-by: Piotr Bielak <pbielak@users.noreply.github.com>